requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
google-genai==0.6.0
importlib-metadata==8.4.0
